Initial version of ATC code.
Everything is still in .h for now, I will fix that. Interfaces with the data structure class (Erica) haven't been fully defined, so that is still not complete. There is a bit of pseudocode still, but I wrote out descriptions so hopefully it makes sense and is close to interfaceable.